+++ crates/mail-uniffi/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Blocking forms of `get_account[s]` and `get_session[s]`
   - Blocking forms of `get_account_state` and `get_session_state`
   - Blocking form of `get_primary_account`
+  - `LabelAs` action set labels to a group of messages 
 
 ## [0.11.56] - 2024-10-03